server/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["CesarMtzV <mc.cesar.art@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.0.3"
Flask-PyMongo = "^2.3.0"
pymongo = {extras = ["srv"], version = "^4.1.1"}
python-dotenv = "^0.20.0"
PyJWT = "^2.3.0"
bcrypt = "^3.2.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
